+++ core-java-modules/core-java-lang-oop-copy/src/main/java/deep/Human.java
@@ -1,0 +1,18 @@
+package deep;
+
+public class Human {
+
+    public Human(String name) {
+        this.name = name;
+    }
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
